refactor(clustering): extract ClusteringService, disentangle clustering from agents
Static analysis, clustering, and the LLM agents were entangled: leaf and
subgraph clustering ran inside the agents via ClusterMethodsMixin, and the
orchestrator reached into agent internals for the separability gate and
the incremental scope inputs. The pipeline now has one clear cut per stage:

  StaticAnalyzer -> StaticAnalysisResults
                 -> ClusteringService -> ClusteringResults
                 -> agents (name/describe/relate the fixed groups)

- clustering/service.py: ClusteringService derives every scope's
  deterministic cluster structure (leaf clusters, scoped call graphs,
  Leiden grouping) for the whole project and per component; it also owns
  the no-callable-structure fail-fast that lived in AbstractionAgent.run.
- clustering/assignment.py: the deterministic post-LLM pass
  (populate_file_methods, assemble_one_component_per_group,
  build_static_relations, ...) as plain functions with explicit inputs;
  agents/cluster_methods_mixin.py is deleted.
- cluster_helpers.py and cluster_relations.py move from static_analyzer/
  to clustering/, which also removes the inverted static_analyzer ->
  agents dependency cluster_relations carried.
- AbstractionAgent.run(clustering) and DetailsAgent.run(component,
  clustering) consume ClusteringResults; DiagramGenerator runs the
  clustering stage explicitly and no longer calls agent-private helpers.

Also fixes a pre-existing gate failure: the pyproject packages test
tripping over locally cloned repos under temp/.
