+++ agents/abstraction_agent.py
@@ -15,7 +15,6 @@ from agents.agent_responses import (
     assign_component_ids,
     assign_relation_ids,
 )
-from agents.cluster_methods_mixin import ClusterMethodsMixin
 from agents.prompts import (
     get_final_analysis_message,
     get_api_surfaces_message,
@@ -31,17 +30,25 @@ from agents.validation import (
     validate_key_entities,
     validate_relations,
 )
+from clustering import ClusteringResults
+from clustering.assignment import (
+    assemble_one_component_per_group,
+    build_scope_cfg_string,
+    build_static_relations,
+    ensure_unique_key_entities,
+    populate_file_methods,
+    resolve_cluster_ids_from_groups,
+)
 from monitoring import trace
-from static_analyzer import StaticAnalysisFatalError
 from static_analyzer.analysis_result import StaticAnalysisResults
-from static_analyzer.cluster_helpers import build_all_cluster_results
-from static_analyzer.constants import Language
 from static_analyzer.graph import ClusterResult
 
 logger = logging.getLogger(__name__)
 
 
-class AbstractionAgent(ClusterMethodsMixin, CodeBoardingAgent):
+class AbstractionAgent(CodeBoardingAgent):
+    """Names and relates the top-level components fixed by the clustering stage."""
+
     def __init__(
         self,
         repo_dir: Path,
@@ -78,21 +85,6 @@ class AbstractionAgent(ClusterMethodsMixin, CodeBoardingAgent):
                 ],
             ),
         }
-
-    @trace
-    def step_clusters_grouping(self, cluster_results: dict[str, ClusterResult]) -> ClusterAnalysis:
-        """Deterministically partition leaf clusters into the top-level component groups.
-
-        Resolution-tuned Leiden picks both the count (modularity peak over
-        ``[5, 8]``) and the membership, so the top-level structure is stable across
-        re-runs — the LLM no longer decides it; it only names them in the
-        final-analysis step.
-        """
-        logger.info(f"[AbstractionAgent] Super-clustering leaf clusters for: {self.project_name}")
-        cfg_graphs = {
-            lang: self.static_analysis.get_cfg(Language(lang)).clustering_networkx() for lang in cluster_results
-        }
-        return self.deterministic_cluster_grouping(cluster_results, cfg_graphs)
 
     @trace
     def step_final_analysis(
@@ -136,7 +128,7 @@ class AbstractionAgent(ClusterMethodsMixin, CodeBoardingAgent):
             validation_context=context,
             max_validation_attempts=3,
         )
-        self.assemble_one_component_per_group(architecture, llm_cluster_analysis, cluster_results)
+        assemble_one_component_per_group(architecture, llm_cluster_analysis, cluster_results)
         return AnalysisInsights(
             description=architecture.description,
             components=architecture.components,
@@ -146,7 +138,7 @@ class AbstractionAgent(ClusterMethodsMixin, CodeBoardingAgent):
     @trace
     def step_api_surfaces(self, analysis: AnalysisInsights) -> ComponentApiSurfaces:
         logger.info(f"[AbstractionAgent] Analyzing component API surfaces for: {self.project_name}")
-        static_call_evidence = self.build_scope_cfg_string(analysis)
+        static_call_evidence = build_scope_cfg_string(analysis, self.static_analysis)
         prompt = self.prompts["api_surfaces"].format(
             component_summaries=analysis.llm_str(),
             static_call_evidence=static_call_evidence,
@@ -162,7 +154,7 @@ class AbstractionAgent(ClusterMethodsMixin, CodeBoardingAgent):
         cluster_results: dict[str, ClusterResult],
     ) -> None:
         logger.info(f"[AbstractionAgent] Discovering component relations for: {self.project_name}")
-        static_call_evidence = self.build_scope_cfg_string(analysis)
+        static_call_evidence = build_scope_cfg_string(analysis, self.static_analysis)
         cfg_graphs = self.static_analysis.available_cfgs()
         self.toolkit.context.cluster_analysis = cluster_analysis
         self.toolkit.context.cluster_results = cluster_results
@@ -188,43 +180,33 @@ class AbstractionAgent(ClusterMethodsMixin, CodeBoardingAgent):
         )
         analysis.components_relations = relation_result.components_relations
         assign_relation_ids(analysis)
-        self.build_static_relations(analysis)
+        build_static_relations(analysis, self.static_analysis)
 
-    def run(self):
-        # Build full cluster results dict for all languages ONCE
-        cluster_results = build_all_cluster_results(self.static_analysis)
+    def run(self, clustering: ClusteringResults) -> AnalysisInsights:
+        """Turn the clustering stage's fixed top-level groups into a named architecture."""
+        cluster_analysis = clustering.cluster_analysis
+        cluster_results = clustering.cluster_results
 
-        # Step 1: Deterministically partition leaf clusters into the top-level groups (Leiden, modularity-peak N)
-        cluster_analysis = self.step_clusters_grouping(cluster_results)
-        if not cluster_analysis.cluster_components:
-            # No leaf clusters means the repo has no callable structure to abstract
-            # (unsupported/empty/no-code). Fail loudly instead of emptying the
-            # architecture and crashing downstream in populate_file_methods.
-            raise StaticAnalysisFatalError(
-                f"No component groups found for {self.project_name}: the static analysis produced "
-                "no callable structure to build an architecture from."
-            )
-
-        # Step 2: Name and describe each fixed group into a component (LLM, one component per group)
+        # Step 1: Name and describe each fixed group into a component (LLM, one component per group)
         analysis = self.step_final_analysis(cluster_analysis, cluster_results)
-        # Step 3: Assign hierarchical component IDs ("1", "2", "3", ...)
+        # Step 2: Assign hierarchical component IDs ("1", "2", "3", ...)
         assign_component_ids(analysis)
-        # Step 4: Resolve cluster IDs deterministically from group names
-        self._resolve_cluster_ids_from_groups(analysis, cluster_analysis)
-        # Step 5: Populate file_methods deterministically from cluster results + orphan assignment
-        self.populate_file_methods(analysis, cluster_results)
+        # Step 3: Resolve cluster IDs deterministically from group names
+        resolve_cluster_ids_from_groups(analysis, cluster_analysis)
+        # Step 4: Populate file_methods deterministically from cluster results + orphan assignment
+        populate_file_methods(analysis, cluster_results, self.repo_dir, self.static_analysis)
 
-        # Step 6: Analyze component API surfaces
+        # Step 5: Analyze component API surfaces
         api_surfaces = self.step_api_surfaces(analysis)
 
-        # Step 7: Discover relations from API surfaces and attach deterministic all_edges
+        # Step 6: Discover relations from API surfaces and attach deterministic all_edges
         self.step_relation_analysis(analysis, api_surfaces, cluster_analysis, cluster_results)
 
-        # Step 8: Fix source code reference lines (resolves reference_file paths for key_entities and key_edges)
+        # Step 7: Fix source code reference lines (resolves reference_file paths for key_entities and key_edges)
         analysis = self.reference_resolver.fix_source_code_reference_lines(analysis)
-        # Step 9: Index relation endpoints after reference resolution
+        # Step 8: Index relation endpoints after reference resolution
         index_relation_endpoints(analysis, self.repo_dir)
-        # Step 10: Ensure unique key entities across components
-        self._ensure_unique_key_entities(analysis)
+        # Step 9: Ensure unique key entities across components
+        ensure_unique_key_entities(analysis)
 
-        return analysis, cluster_results
+        return analysis
